Update: Use long for usage fields, rename of age to recency, correct printing of negation, fix of answer condition in example which led to correctly asnwered questions being miscounted as failed

+++ src/Usage.h
@@ -13,9 +13,9 @@
 //--------------//
 typedef struct {
     //use_count, how often it was used in total
-    int useCount;
+    long useCount;
     //age, how many cycles ago it was last used
-    int lastUsed;
+    long lastUsed;
 } Usage;
 
 //Methods//
